feat(courses): add access guard and layout component for courses
Extract common prefetch logic and access control into Courses.layout component
Mark global settings endpoint  public in API
