+++ apps/web/app/modules/Courses/Courses.page.tsx
@@ -3,20 +3,11 @@ import { useReducer } from "react";
 import { useTranslation } from "react-i18next";
 import { match } from "ts-pattern";
 
-import {
-  availableCoursesQueryOptions,
-  currentUserQueryOptions,
-  studentCoursesQueryOptions,
-  useCategoriesSuspense,
-} from "~/api/queries";
+import { useCategoriesSuspense } from "~/api/queries";
 import { useAvailableCourses } from "~/api/queries/useAvailableCourses";
-import { categoriesQueryOptions } from "~/api/queries/useCategories";
-import { allCoursesQueryOptions } from "~/api/queries/useCourses";
-import { queryClient } from "~/api/queryClient";
 import { ButtonGroup } from "~/components/ButtonGroup/ButtonGroup";
 import { Icon } from "~/components/Icon";
 import { PageWrapper } from "~/components/PageWrapper";
-import { USER_ROLE } from "~/config/userRoles";
 import { useUserRole } from "~/hooks/useUserRole";
 import { cn } from "~/lib/utils";
 import { useLayoutsStore } from "~/modules/common/Layout/LayoutsStore";
@@ -31,39 +22,11 @@ import { StudentsCurses } from "~/modules/Courses/components/StudentsCurses";
 import { DashboardIcon, HamburgerIcon } from "~/modules/icons/icons";
 import { createSortOptions, type SortOption } from "~/types/sorting";
 
-import type { MetaFunction } from "@remix-run/node";
-
-export const meta: MetaFunction = () => {
-  return [{ title: "Courses" }, { name: "description", content: "Courses" }];
-};
-
-const prefetchQueriesForUser = async (userRole: string | undefined) => {
-  await queryClient.prefetchQuery(categoriesQueryOptions());
-
-  return match(userRole)
-    .with(USER_ROLE.admin, USER_ROLE.contentCreator, async () => {
-      await queryClient.prefetchQuery(allCoursesQueryOptions());
-    })
-    .with(USER_ROLE.student, async () => {
-      await queryClient.prefetchQuery(availableCoursesQueryOptions());
-      await queryClient.prefetchQuery(studentCoursesQueryOptions());
-    })
-    .otherwise(async () => {
-      await queryClient.prefetchQuery(availableCoursesQueryOptions());
-    });
-};
-
-export const clientLoader = async () => {
-  const currentUser = await queryClient.ensureQueryData(currentUserQueryOptions);
-  const userRole = currentUser?.data.role;
-
-  await prefetchQueriesForUser(userRole);
-
-  return null;
-};
+import { CoursesAccessGuard } from "./Courses.layout";
 
 export default function CoursesPage() {
   const { isStudent } = useUserRole();
+  const { t } = useTranslation();
 
   type State = {
     searchTitle: string | undefined;
@@ -76,7 +39,6 @@ export default function CoursesPage() {
     | { type: "SET_SORT"; payload: string | undefined }
     | { type: "SET_CATEGORY"; payload: string | undefined };
 
-  const { t } = useTranslation();
   function reducer(state: State, action: Action): State {
     return match<Action, State>(action)
       .with({ type: "SET_SEARCH_TITLE" }, ({ payload }) => ({
@@ -148,82 +110,84 @@ export default function CoursesPage() {
   };
 
   return (
-    <PageWrapper>
-      <div className="flex h-auto flex-1 flex-col gap-y-12">
-        {isStudent && <StudentsCurses />}
-        <div className="flex flex-col">
-          <div className="flex flex-col lg:p-0">
-            <h4 className="pb-1 text-2xl font-bold leading-10 text-neutral-950">
-              {t("studentCoursesView.availableCourses.header")}
-            </h4>
-            <p className="text-lg leading-7 text-neutral-800">
-              {t("studentCoursesView.availableCourses.subHeader")}
-            </p>
-            <div className="flex items-center justify-between gap-2">
-              <SearchFilter
-                filters={filterConfig}
-                values={{
-                  title: state.searchTitle,
-                  category: state.category,
-                  sort: state.sort,
-                }}
-                onChange={handleFilterChange}
-                isLoading={isCategoriesLoading}
-              />
-              <ButtonGroup
-                className="sr-only lg:not-sr-only"
-                buttons={[
-                  {
-                    children: <DashboardIcon />,
-                    isActive: courseListLayout === "card",
-                    onClick: () => setCourseListLayout("card"),
-                  },
-                  {
-                    children: <HamburgerIcon />,
-                    isActive: courseListLayout === "table",
-                    onClick: () => setCourseListLayout("table"),
-                  },
-                ]}
-              />
+    <CoursesAccessGuard>
+      <PageWrapper>
+        <div className="flex h-auto flex-1 flex-col gap-y-12">
+          {isStudent && <StudentsCurses />}
+          <div className="flex flex-col">
+            <div className="flex flex-col lg:p-0">
+              <h4 className="pb-1 text-2xl font-bold leading-10 text-neutral-950">
+                {t("studentCoursesView.availableCourses.header")}
+              </h4>
+              <p className="text-lg leading-7 text-neutral-800">
+                {t("studentCoursesView.availableCourses.subHeader")}
+              </p>
+              <div className="flex items-center justify-between gap-2">
+                <SearchFilter
+                  filters={filterConfig}
+                  values={{
+                    title: state.searchTitle,
+                    category: state.category,
+                    sort: state.sort,
+                  }}
+                  onChange={handleFilterChange}
+                  isLoading={isCategoriesLoading}
+                />
+                <ButtonGroup
+                  className="sr-only lg:not-sr-only"
+                  buttons={[
+                    {
+                      children: <DashboardIcon />,
+                      isActive: courseListLayout === "card",
+                      onClick: () => setCourseListLayout("card"),
+                    },
+                    {
+                      children: <HamburgerIcon />,
+                      isActive: courseListLayout === "table",
+                      onClick: () => setCourseListLayout("table"),
+                    },
+                  ]}
+                />
+              </div>
+            </div>
+            <div
+              data-testid="unenrolled-courses"
+              className={cn("gap-6 rounded-lg drop-shadow-primary lg:bg-white lg:p-8", {
+                "flex flex-wrap": courseListLayout === "card",
+                block: courseListLayout === "table",
+              })}
+            >
+              {userAvailableCourses && !isEmpty(userAvailableCourses) && (
+                <CourseList
+                  availableCourses={userAvailableCourses}
+                  courseListLayout={courseListLayout}
+                />
+              )}
+              {!userAvailableCourses ||
+                (isEmpty(userAvailableCourses) && (
+                  <div className="col-span-3 flex gap-8">
+                    <div>
+                      <Icon name="EmptyCourse" className="mr-2 text-neutral-900" />
+                    </div>
+                    <div className="flex flex-col justify-center gap-2">
+                      <p className="text-lg font-bold leading-5 text-neutral-950">
+                        {t("studentCoursesView.other.cannotFindCourses")}
+                      </p>
+                      <p className="text-base font-normal leading-6 text-neutral-800">
+                        {t("studentCoursesView.other.changeSearchCriteria")}
+                      </p>
+                    </div>
+                  </div>
+                ))}
+              {isAvailableCoursesLoading && (
+                <div className="flex h-full items-center justify-center">
+                  <Loader />
+                </div>
+              )}
             </div>
           </div>
-          <div
-            data-testid="unenrolled-courses"
-            className={cn("gap-6 rounded-lg drop-shadow-primary lg:bg-white lg:p-8", {
-              "flex flex-wrap": courseListLayout === "card",
-              block: courseListLayout === "table",
-            })}
-          >
-            {userAvailableCourses && !isEmpty(userAvailableCourses) && (
-              <CourseList
-                availableCourses={userAvailableCourses}
-                courseListLayout={courseListLayout}
-              />
-            )}
-            {!userAvailableCourses ||
-              (isEmpty(userAvailableCourses) && (
-                <div className="col-span-3 flex gap-8">
-                  <div>
-                    <Icon name="EmptyCourse" className="mr-2 text-neutral-900" />
-                  </div>
-                  <div className="flex flex-col justify-center gap-2">
-                    <p className="text-lg font-bold leading-5 text-neutral-950">
-                      {t("studentCoursesView.other.cannotFindCourses")}
-                    </p>
-                    <p className="text-base font-normal leading-6 text-neutral-800">
-                      {t("studentCoursesView.other.changeSearchCriteria")}
-                    </p>
-                  </div>
-                </div>
-              ))}
-            {isAvailableCoursesLoading && (
-              <div className="flex h-full items-center justify-center">
-                <Loader />
-              </div>
-            )}
-          </div>
         </div>
-      </div>
-    </PageWrapper>
+      </PageWrapper>
+    </CoursesAccessGuard>
   );
 }
